fix: bound libsodium install in NLG pod and give it its own test step (#5988)
