+++ src/commands/rapid-fire.ts
@@ -3,6 +3,7 @@
 import {Listr} from 'listr2';
 import {SoloErrors} from '../core/errors/solo-errors.js';
 import * as constants from '../core/constants.js';
+import {NetworkLoadGeneratorLibraries} from '../core/network-load-generator-libraries.js';
 import {BaseCommand} from './base.js';
 import {Flags as flags} from './flags.js';
 import {type AnyListrContext, type ArgvStruct} from '../types/aliases.js';
@@ -244,10 +245,7 @@ export class RapidFireCommand extends BaseCommand {
                   pod.podReference,
                   constants.NETWORK_LOAD_GENERATOR_CONTAINER,
                 );
-                const container: Container = k8Containers.readByRef(containerReference);
-                await container.execContainer('apt-get update -qq');
-                await container.execContainer('apt-get install -y libsodium23');
-                await container.execContainer('apt-get clean -qq');
+                await NetworkLoadGeneratorLibraries.install(k8Containers.readByRef(containerReference));
               }
             },
           },
